move get all checkins to /users/:key, add delete check-in api
