+++ api/index.js
@@ -3,5 +3,6 @@ const express = require("express");
 const router = express.Router();
 
 router.use("/check-ins", require("./check-ins"));
+router.use("/users", require("./users"));
 
 module.exports = router;
